\name{draw_elev3Drgl_OCN}
\alias{draw_elev3Drgl_OCN}
\title{Plot 3D map of elevation generated by an OCN via rgl rendering}
\usage{
draw_elev3Drgl_OCN(OCN, coarse_grain = c(1, 1), chooseCM = FALSE,
  add_colorbar = FALSE, draw_river = FALSE, A_thr_draw = 0.002 *
  OCN$dimX * OCN$dimY * OCN$cellsize^2, river_color = "#00CCFF", ...)
}
\arguments{
\item{OCN}{List as produced by \code{\link{landscape_OCN}}.}

\item{coarse_grain}{2x1 vector (only effective if \code{chooseCM = FALSE}). For aesthetic purposes, the elevation map can be coarse-grained into a \code{OCN$dimX/coarse_grain[1]}-by-\code{OCN$dimX/coarse_grain[2]} 
domain, where each cell's elevation is the average of elevations of the corresponding \code{coarse_grain[1]}-by-\code{coarse_grain[2]} cells of the original elevation field. 
\code{coarse_grain[1]} and \code{coarse_grain[2]} must be divisors of \code{OCN$dimX} and \code{OCN$dimY}, respectively.}

\item{chooseCM}{Index of catchment to display (only effective if \code{OCN$N_outlet > 1}). It can be a logical, or a scalar within \code{1:length(OCN$N_outlet)}. 
If \code{TRUE}, the catchment with largest area is displayed. Note that, if the size of the chosen catchment is too small (e.g. \code{OCN$CM$A[chooseCM] < 5*OCN$cellsize^2}), an error might occur due to failure in triangulation.}

\item{add_colorbar}{If \code{TRUE}, add colorbar to the plot.}

\item{draw_river}{If \code{TRUE}, draw the OCN on top of the elevation field.}

\item{A_thr_draw}{Threshold drainage area value used to display the network.}

\item{river_color}{Color used to plot the river.}

\item{...}{Further parameters passed to function \code{\link{persp3d}}.}
}
\value{
No output is returned.
}
\description{
Function that plots the 3D elevation map generated by an OCN.
}
\details{
This function makes use of the \code{\link{rgl}} rendering system. To export the figure in raster format, use \code{\link{rgl.snapshot}}. 
To export in vectorial format, use \code{\link{rgl.postscript}} (but note that this might produce rendering issues, see \code{\link{rgl}} for details).
The function will attempt at drawing a contour of the plotted entity (i.e. the lattice or a catchment, depending on \code{chooseCM}) at null elevation, and drawing polygons connecting this contour with the lattice/catchment contour at the real elevation. If \code{chooseCM != FALSE}, this might result in errors owing to failure of \code{\link{polygon3d}} in triangulating the polygons.  
}
\examples{
\dontrun{draw_elev3Drgl_OCN(landscape_OCN(OCN_20))}
}
