\name{draw_subcatchments_OCN}
\alias{draw_subcatchments_OCN}
\title{Draw subcatchment map from an Optimal Channel Network}
\usage{
draw_subcatchments_OCN(OCN, draw_river = TRUE)
}
\arguments{
\item{OCN}{List as produced by \code{\link{aggregate_OCN}}.}

\item{draw_river}{if \code{TRUE}, draw the OCN on top of the elevation field.}
}
\value{
No output is returned.
}
\description{
Function that draws a map of subcatchments generated by the aggregation process on the OCN.
}
\examples{
draw_subcatchments_OCN(aggregate_OCN(landscape_OCN(OCN_20)))
}
