\docType{data}
\name{flowDir}
\alias{flowDir}}
\title{Example of initial flow direction matrix to be used as input in \code{create_general_contour_OCN}.}
\format{A matrix. See \code{\link{create_general_contour_OCN}} documentation for details.}
\usage{
data(flowDir)
}
\description{
An arbitrary initial flow direction matrix.
}
\keyword{datasets}
