\docType{data}
\name{OCN_20}
\alias{OCN_20}
\title{Example of small OCN}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_20)
}
\description{
A network built on a 20x20 lattice obtained by executing \code{set.seed(1); create_OCN(20,20)}.
}
\keyword{datasets}
