\docType{data}
\name{OCN_250_cold}
\alias{OCN_250_cold}
\title{Example of single-outlet OCN}
\format{A \code{river} object. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_250_cold)
}
\description{
A network built on a 250x250 lattice obtained by executing \code{set.seed(2); create_OCN(250, 250, coolingRate = 10)}.
}
\keyword{datasets}
