\docType{data}
\name{OCN_250_PB}
\alias{OCN_250_PB}
\title{Example of single-outlet OCN with periodic boundaries}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_250_PB)
}
\description{
A network built on a 250x250 lattice obtained by executing \code{set.seed(2); create_OCN(250, 250, periodicBoundaries = TRUE)}.
}
\keyword{datasets}
