\docType{data}
\name{OCN_250_V}
\alias{OCN_250_V}
\title{Example of single-outlet OCN}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_250_V)
}
\description{
A network built on a 250x250 lattice obtained by executing \code{set.seed(2); create_OCN(250, 250, typeInitialState = "V")}.
}
\keyword{datasets}
