\docType{data}
\name{OCN_250_V_cold}
\alias{OCN_250_V_cold}
\title{Example of single-outlet OCN}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_250_V_cold)
}
\description{
A network built on a 250x250 lattice with \code{type_initialstate = "V"}, \code{CoolingRate = 200}. Other parameters are set to their default values.
}
\keyword{datasets}
