\docType{data}
\name{OCN_300_7out}
\alias{OCN_300_7out}
\title{Example of multiple-outlet OCN}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_300_7out)
}
\description{
A network built on a 300x300 lattice obtained by executing \code{set.seed(5); create_OCN(300,300,N_outlet = 7,cellsize = 50)}.
}
\keyword{datasets}
