\docType{data}
\name{OCN_300_diag}
\alias{OCN_300_diag}
\title{Example of single-outlet OCN}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_300_diag)
}
\description{
A network built on a 300x300 lattice obtained by executing \code{set.seed(4); create_OCN(300, 300, outletPos = 1, typeInitialState = "V", cellsize = 50)}.
}
\keyword{datasets}
