\docType{data}
\name{OCN_4}
\alias{OCN_4}
\title{Example of small OCN}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_4)
}
\description{
A network built on a 4x4 lattice for illustrative purposes.
}
\details{
Despite its name, this network is not an OCN: indeed, it has been generated manually and not via \code{\link{create_OCN}}.
}
\keyword{datasets}
