\docType{data}
\name{OCN_400_Allout}
\alias{OCN_400_Allout}
\title{Example of OCN with all perimetric pixels as outlets}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_400_Allout)
}
\description{
A network built on a 400x400 lattice obtained by executing \code{set.seed(8); create_OCN(400, 400, nOutlet = "All", cellsize = 50)}.
}
\keyword{datasets}
