\docType{data}
\name{OCN_rect1}
\alias{OCN_rect1}
\title{Example of single-outlet OCN}
\format{A list. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_rect1)
}
\description{
A network built on a 450x150 lattice obtained by executing \code{set.seed(3); create_OCN(450,150)}.
}
\keyword{datasets}
