\docType{data}
\name{OCN_rect2}
\alias{OCN_rect2}
\title{Example of single-outlet OCN}
\format{A \code{river} object. See \code{\link{create_OCN}} documentation for details.}
\usage{
data(OCN_rect2)
}
\description{
A network built on a 150x450 lattice obtained by executing \code{set.seed(3); create_OCN(150, 450)}.
}
\keyword{datasets}
