\name{OCN_to_SSN}
\alias{OCN_to_SSN}
\title{Transform OCN into SSN object (disabled)}
\usage{
OCN_to_SSN(OCN, ...)
}
\arguments{
\item{OCN}{A \code{river} object.}

\item{...}{Further arguments.}

}
\value{
No output is produced.
}
\description{
In \code{OCNet} v1.2.0, this function is disabled following the archiving 
of the \code{SSN} package from CRAN (due to the retirement of \code{rgdal}). 
A new \code{OCN_to_SSN} function will be included in the next release of \code{OCNet}.
}
\examples{
OCN_to_SSN(OCN)
}
