\docType{package}
\name{OCNet-package}
\alias{OCNet-package}
\alias{OCNet}
\title{Create and analyze Optimal Channel Networks.}
\description{
A package that allows the generation and analysis of synthetic river network analogues, called Optimal Channel Networks (OCNs).
}

\section{References}{


Rinaldo, A., Rigon, R., Banavar, J. R., Maritan, A., & Rodriguez-Iturbe, I. (2014). Evolution and selection of river networks: Statics, dynamics, and complexity. Proceedings of the National Academy of Sciences of the United States of America, 111(7), 2417-2424. doi:10.1073/pnas.1322700111

Carraro, L., Bertuzzo, E., Fronhofer, E. A., Furrer, R., Gounand, I., Rinaldo, A., & Altermatt, F. (2020). Generation and application of river network analogues for use in ecology and evolution. Ecology and Evolution, 10(14), 7537-7550. doi:10.1002/ece3.6479

Carraro, L., & Altermatt, F. (2022). Optimal channel networks accurately model ecologically-relevant geomorphological features of branching river networks. Communications Earth and Environment, 3(1) doi:10.1038/s43247-022-00454-1
}

\seealso{
\href{../doc/OCNet.html}{\code{vignette("OCNet")}}
}
\author{
Luca Carraro (\email{luca.carraro@hotmail.it}), Florian Altermatt, Emanuel A. Fronhofer, Reinhard Furrer, Isabelle Gounand, Andrea Rinaldo, Enrico Bertuzzo
}
