\name{permuteAddSolve}
\alias{permuteAddSolve}
\title{Permute add solve}
\usage{
permuteAddSolve(Wt, perm, n, nNodes, expEnergy)
}
\arguments{
\item{Wt}{transposed adjacency matrix.}

\item{perm}{permutation vector.}

\item{n}{should be equal to 1.}

\item{nNodes}{number of nodes.}

\item{expEnergy}{exponent of energy.}
}
\value{
A list.
}
\description{
Function that permutes, adds and solves.
}
