\name{draw_elev3D_OCN}
\alias{draw_elev3D_OCN}
\title{Plot 3D map of elevation generated by an OCN}
\usage{
draw_elev3D_OCN(OCN, ColPalette = terrain.colors(1000, alpha = 1),
  add_colorbar = TRUE, draw_river = TRUE, A_thr_draw = 0.002 *
  OCN$dimX * OCN$dimY * OCN$cellsize^2, river_color = "#00CCFF")
}
\arguments{
\item{OCN}{List as produced by \code{\link{landscape_OCN}}.}

\item{ColPalette}{Color palette used for the plot.}

\item{add_colorbar}{If \code{TRUE}, add colorbar to the plot.}

\item{draw_river}{If \code{TRUE}, draw the OCN on top of the elevation field.}

\item{A_thr_draw}{Threshold drainage area value used to display the network.}

\item{river_color}{Color used to plot the river.}
}
\value{
No output is returned.
}
\description{
Function that plots the 3D elevation map generated by an OCN.
}
\examples{
draw_elev3D_OCN(landscape_OCN(OCN_20))
}
