\docType{package}
\name{OCNet-package}
\alias{OCNet-package}
\alias{OCNet}
\title{Create and analyze Optimal Channel Networks.}
\description{
A package that allows the generation and analysis of synthetic river network analogues, called Optimal Channel Networks (OCNs).
}

\section{References}{


Rinaldo, A., Rigon, R., Banavar, J. R., Maritan, A., & Rodriguez-Iturbe, I. (2014). Evolution and selection of river networks: Statics, dynamics, and complexity. Proceedings of the National Academy of Sciences of the United States of America, 111(7), 2417-2424. doi:10.1073/pnas.1322700111

Carraro et al.
}

\seealso{
\href{../doc/OCNet.html}{\code{vignette("OCNet")}}
}
\author{
Luca Carraro (\email{luca.carraro@hotmail.it}), Florian Altermatt, Emanuel A. Fronhofer, Reinhard Furrer, Isabelle Gounand, Andrea Rinaldo, Enrico Bertuzzo
}
